클럽 운영자 조회(멀티건) 카운트 단위테스트

+++ src/main/java/egovframework/com/cop/com/service/impl/EgovUserInfManageDAO.java
@@ -136,12 +136,12 @@ public class EgovUserInfManageDAO extends EgovComAbstractDAO {
 	 * 동호회 운영자 목록에 대한 전체 건수를 조회한다.
 	 *
 	 * @param userVO
-	 * @return
+	 * @return int clubOprtrsCnt = egovUserInfManageDAO.selectClubOprtrListCnt(userVO);
 	 * @throws Exception
 	 */
 	@Deprecated
 	public int selectClubOprtrListCnt(UserInfVO userVO) throws Exception {
-		return (Integer) selectOne("EgovUserInfManageDAO.selectClubOprtrListCnt", userVO);
+		return selectOne("EgovUserInfManageDAO.selectClubOprtrListCnt", userVO);
 	}
 
 	/**
